control d

+++ srcs/1_lineeditor/control_keys.c
@@ -6,7 +6,7 @@
 /*   By: pespalie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/18 08:49:38 by pespalie          #+#    #+#             */
-/*   Updated: 2018/12/18 09:44:35 by pespalie         ###   ########.fr       */
+/*   Updated: 2018/12/18 10:14:37 by pespalie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,19 +36,18 @@ char	*delete_char(int **index, char *line, int len)
 char	*control_d(char *line, int *index)
 {
 	char	*ret;
-	int		len;
+	size_t	len;
 	t_data	*data;
 
 	ret = NULL;
 	data = get_data();
 	data->eof = 1;
-	len = (int)ft_strlen(line);
+	len = ft_strlen(line);
 	if (len < 2)
 		len = 2;
 	if (ft_strcmp(line, "\0") == 0)
 		exit(1);
-	if (*index == len)
-//CLEAR LINE + DISPLAY PROMPT
+	if (*index == (int)len)
 		return (line);
 	ret = delete_char(&index, line, len);
 	free(line);
